+++ src/translation/services/Translator/drivers/FileDriver.js
@@ -23,7 +23,7 @@ class FileDriver extends Driver {
 	 * @type {Array<string>}
 	 */
 	static get dependencies() {
-		return (super.dependencies || []).concat(['app', 'file']);
+		return (super.dependencies || []).concat(['app', 'file', 'helper.string']);
 	}
 
 	/**
@@ -91,6 +91,12 @@ class FileDriver extends Driver {
 		if (!__(this).get('loaded')) {
 			if (this.file.exists(this.folder)) {
 				const translations = this.file.loadRecursivelyInFolder(this.folder);
+				Object.entries(translations).forEach(([key, value]) => {
+					if (key.includes('/')) {
+						delete translations[key];
+						dot.str(this.stringHelper.dot(key), value, translations);
+					}
+				});
 				this.addTranslations(translations);
 				__(this).set('loaded', true);
 			}
@@ -203,6 +209,15 @@ class FileDriver extends Driver {
 	 */
 	get defaultNamespace() {
 		return 'translations';
+	}
+
+	/**
+	 * String helper.
+	 *
+	 * @type {support.helpers.StringHelper}
+	 */
+	get stringHelper() {
+		return this.helperString;
 	}
 
 }